+++ treipigmei-chess-ai/src/board/Move.java
@@ -2,7 +2,7 @@ package board;
 
 public class Move {
     private String move;
-    private byte [] pos;
+    private int [] pos;
 
     /*
      * TODO De testat daca mutarea are mai mult de 4 caractere lungime
@@ -14,7 +14,7 @@ public class Move {
      */
     public Move(String s) {
         this.move = new String(s);
-        this.pos = new byte[4];
+        this.pos = new int[4];
         decodeMove();
     }
     
@@ -22,8 +22,8 @@ public class Move {
      * 
      * @return Source position
      */
-    public byte [] getFrom() {
-        byte [] aux = new byte[2];
+    public int [] getFrom() {
+        int [] aux = new int[2];
         aux[0] = this.pos[0];
         aux[1] = this.pos[1];
         
@@ -34,8 +34,8 @@ public class Move {
      * 
      * @return Destination position
      */
-    public byte [] getTo() {
-        byte [] aux = new byte[2];
+    public int [] getTo() {
+        int [] aux = new int[2];
         aux[0] = this.pos[2];
         aux[1] = this.pos[3];
         
@@ -48,10 +48,10 @@ public class Move {
      */
      private void decodeMove() {
     	 
-         pos[0] = (byte)(8 - (move.charAt(1) - '0'));
-         pos[1] = (byte)(move.charAt(0) - 'a');
-         pos[2] = (byte)(8 - (move.charAt(3) - '0'));
-         pos[3] = (byte)(move.charAt(2) - 'a');
+         pos[0] = (8 - (move.charAt(1) - '0'));
+         pos[1] = (move.charAt(0) - 'a');
+         pos[2] = (8 - (move.charAt(3) - '0'));
+         pos[3] = (move.charAt(2) - 'a');
     }
     
 }
