+++ treipigmei-chess-ai/src/piece/BlackPawn.java
@@ -4,10 +4,10 @@ import helpers.Flags.Colour;
 
 public class BlackPawn extends Piece {
     
-    public BlackPawn(Colour color, byte[] position) {
+    public BlackPawn(Colour color, int[] position) {
         super(color, position);
-        x = new byte[]{0, -1, 1};
-    	y = new byte[]{1, 1, 1};
+        x = new int[]{0, -1, 1};
+    	y = new int[]{1, 1, 1};
     }
