+++ treipigmei-chess-ai/src/piece/King.java
@@ -4,10 +4,10 @@ import helpers.Flags.Colour;
 
 public class King extends Piece {
     
-    public King(Colour color, byte[] position) {
+    public King(Colour color, int[] position) {
         super(color, position);
-        x = new byte[] {-1, -1, 1, 1, -1, 1, 0, 0};
-    	y = new byte[] {1, -1, -1, 1, 0, 0, -1, 1};
+        x = new int[] {-1, -1, 1, 1, -1, 1, 0, 0};
+    	y = new int[] {1, -1, -1, 1, 0, 0, -1, 1};
     }
