+++ treipigmei-chess-ai/src/piece/Rook.java
@@ -4,10 +4,10 @@ import helpers.Flags.Colour;
 
 public class Rook extends Piece {
     
-    public Rook(Colour color, byte[] position) {
+    public Rook(Colour color, int[] position) {
         super(color, position);
-        x = new byte[]{-1, 1, 0, 0};
-        y = new byte[]{0, 0, -1, 1};
+        x = new int[]{-1, 1, 0, 0};
+        y = new int[]{0, 0, -1, 1};
     }
 
 }
